+++ primo2/tests/Inverence_test.py
@@ -21,7 +21,7 @@
 
 import unittest
 import numpy as np
-from primo2.network import BayesianNetwork
+from primo2.networks import BayesianNetwork
 from primo2.nodes import DiscreteNode
 from primo2.io import XMLBIFParser
 from primo2.inference.order import Orderer
